Change some variables to LogDensity.

+++ src/core/dag/DagNode.cpp
@@ -246,11 +246,11 @@ void DagNode::executeMethod(const std::string &n, const std::vector<const DagNod
 
     if ( n == "lnProbability" )
     {
-        rv = const_cast<DagNode *>(this)->getLnProbability();
+        rv = (double)const_cast<DagNode *>(this)->getLnProbability();
     }
     else if ( n == "probability" )
     {
-        rv = std::exp(const_cast<DagNode *>(this)->getLnProbability());
+        rv = exp(const_cast<DagNode *>(this)->getLnProbability());
     }
     else
     {
